Abstract class and methods and Concrete class and methods.
This is the example of Abstract and concrete.
Abstract is a key word Which can be use with a class and methods. Abstract class can't be overridden. Abstract class can store both Abstract method and concrete method.
Concrete class and methods
    If a class is not a declared using abstract key word use referred as concreate class are non-abstract class.
concrete class can store only concrete method, concreate method as both declaration and implementation